apps/corporate-website/api/backend/requirements.txt
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.29.0,<0.31.0
motor>=3.4.0,<4.0.0
python-dotenv>=1.0.1,<2.0.0
pydantic>=2.7.0,<3.0.0
